@@ -1,48 +1,37 @@
 {
   "name": "Hrguru",
-  "private": true,
   "devDependencies": {
-    "babel-eslint": "^4.1.1",
-    "eslint": "^0.24.1",
-    "eslint-plugin-react": "^3.3.1",
-    "expect": "^1.9.0",
-    "geckodriver": "^1.3.0",
-    "karma": "^0.13.9",
-    "karma-chrome-launcher": "^0.2.0",
+    "babel-eslint": "^4.1.3",
+    "eslint": "^1.7.3",
+    "eslint-plugin-react": "^3.6.3",
+    "expect": "^1.12.2",
+    "karma": "^0.13.13",
+    "karma-chrome-launcher": "^0.2.1",
     "karma-mocha": "^0.2.0",
     "karma-sinon": "^1.0.4",
-    "karma-sourcemap-loader": "^0.3.5",
+    "karma-sourcemap-loader": "^0.3.6",
     "karma-webpack": "^1.7.0",
     "react-hot-loader": "^1.3.0",
     "webpack-dev-server": "^1.12.1",
     "webpack-notifier": "^1.2.1"
   },
   "dependencies": {
-    "alt": "0.17.9",
-    "babel-core": "^5.8.19",
+    "babel-core": "^5.8.25",
     "babel-loader": "^5.3.2",
-    "classnames": "^2.2.0",
-    "css-loader": "0.22.0",
     "extract-text-webpack-plugin": "^0.8.2",
-    "lodash": "^4.17.2",
-    "moment": "^2.10.6",
-    "node-sass": "^3.2.0",
-    "react": "^0.14.2",
-    "react-bootstrap": "0.27.1",
-    "react-bootstrap-daterangepicker": "1.0.2",
-    "react-copy-to-clipboard": "^3.0.4",
-    "react-day-picker": "^1.1.5",
-    "react-dom": "0.14.2",
-    "react-modal": "0.6.1",
-    "react-select": "0.9.1",
+    "node-sass": "^3.3.3",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-router": "1.0.0-rc3",
     "react-tools": "*",
-    "sass-loader": "^2.0.1",
+    "sass-loader": "^3.0.0",
     "webpack": "^1.12.1"
   },
   "scripts": {
     "test": "karma start",
     "start": "webpack -w --config webpack/dev.config.js",
-    "build": "webpack --optimize-occurrence-order --config webpack/production.config.js"
+    "start-hot-dev": "webpack-dev-server --hot --inline --config webpack/hot-dev.config.js",
+    "build": "webpack -p --config webpack/production.config.js"
   },
   "license": "",
   "engines": {
