@@ -1,20 +1,8 @@
-window.React = require('react');
-
+import React from 'react';
 import ReactDOM from 'react-dom';
+window.React = React;
 window.ReactDOM = ReactDOM;
 
-import Statistics from './components/statistics/statistics';
-import Teams from './components/teams/teams';
-import Projects from './components/projects/projects';
-import Users from './components/users/users';
-import UserSkillRates from './components/skills/user_skill_rates';
-import UserSkillHistory from './components/skills/user_skill_history';
-import Scheduling from './components/scheduling/scheduling';
+import HelloWorld from './components/hello-world';
+registerComponent('HelloWorld', HelloWorld);
 
-registerComponent('statistics', Statistics);
-registerComponent('teams', Teams);
-registerComponent('projects', Projects);
-registerComponent('users', Users);
-registerComponent('user_skill_rates', UserSkillRates);
-registerComponent('user_skill_history', UserSkillHistory);
-registerComponent('scheduling', Scheduling);
